add decimal operation feature

+++ sundanase_interpreter.py
@@ -7,14 +7,14 @@ class BasicExecute:
     def __init__(self, tree, env):
         self.env = env
         result = self.walkTree(tree)
-        if result is not None and isinstance(result, int):
+        if result is not None and isinstance(result, float):
             print(result)
         if isinstance(result, str) and result[0] == '"':
             print(result)
 
     def walkTree(self, node):
 
-        if isinstance(node, int):
+        if isinstance(node, float):
             return node
         if isinstance(node, str):
             return node
